# docs/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "TalentGift Rest API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module.",
    "version": "1.0.0"
  },
  "host": "localhost:5000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/resetPass": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "type": "string"
          },
          {
            "name": "password",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/individualReport": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "uid",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/managerReport": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "uid",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/organizationReport": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "oid",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/criticalPositions": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "oid",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/expectations": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "oid",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/password/{uid}": {
      "patch": {
        "description": "",
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
  }
}